add User view implementation;

See merge request foxstudent104853/universitycms!7

+++ src/main/java/ua/foxminded/pskn/universitycms/dto/FacultyDTO.java
@@ -16,6 +16,6 @@ public class FacultyDTO {
 
     private String facultyName;
 
-    private int universityId;
+    private Long universityId;
 
 }
